add snwprintf_s, vsnwprintf_s, swprintf_s
add special code with a temp. stack buffer
to detect ENOSPC with the untruncated versions.

also improve vswprintf_s:
for swprintf_s, vswprintf_s add manual ESNOSPC checks
on -1, trying larger with a larger buffer, to check
if dmax was too small, or any other (e.g. encoding) error.

for the truncating sn versions:
similar: manually add truncation, as the original wide API is
pretty broken.

src/Makefile: add %.c.i target with -E preprocessed output
add DSYMUTIL helper for rhapsody/darwin for the test executables.

revert CFLAGS removal, we really need the old logic.

+++ src/ignore_handler_s.c
@@ -56,7 +56,7 @@ void ignore_handler_s(const char * restrict msg, void * restrict ptr, errno_t er
 	(void)msg;
 	(void)error;
 #endif
-	sldebug_printf("IGNORE CONSTRAINT HANDLER: (%u) %s\n", error,
+	sldebug_printf("IGNORE CONSTRAINT HANDLER: (%d) %s\n", (int)error,
 		       (msg) ? msg : "Null message");
 	return;
 }
